refactor: Separate calendar code (#8)
